refact: improve perf for cc algorithm (#49)

+++ computer-algorithm/src/main/java/com/baidu/hugegraph/computer/algorithm/community/trianglecount/TriangleCountOutput.java
@@ -33,7 +33,7 @@ public class TriangleCountOutput extends HugeOutput {
     @Override
     public void prepareSchema() {
         this.client().schema().propertyKey(this.name())
-                     .asLong()
+                     .asInt()
                      .writeType(WriteType.OLAP_RANGE)
                      .ifNotExist()
                      .create();
@@ -43,9 +43,9 @@ public class TriangleCountOutput extends HugeOutput {
     public com.baidu.hugegraph.structure.graph.Vertex constructHugeVertex(
                                                       Vertex vertex) {
         com.baidu.hugegraph.structure.graph.Vertex hugeVertex =
-                new com.baidu.hugegraph.structure.graph.Vertex(null);
+                  new com.baidu.hugegraph.structure.graph.Vertex(null);
         hugeVertex.id(vertex.id().asObject());
-        long value = ((TriangleCountValue) vertex.value()).count();
+        int value = ((TriangleCountValue) vertex.value()).count();
         hugeVertex.property(this.name(), value);
         return hugeVertex;
     }
